Move TableType string methods to string module
